增加列表、详情显示

+++ backend/src/main/java/com/dcits/ms/database/InitDatabase.java
@@ -5,8 +5,11 @@ import com.dcits.ms.service.BaseProjectService;
 import com.dcits.ms.service.DepartmentService;
 import com.dcits.ms.service.ProductService;
 import com.dcits.ms.service.UserService;
+import com.dcits.ms.util.BusiUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+
+import java.util.List;
 
 
 @Component
@@ -23,31 +26,35 @@ public class InitDatabase {
 
     String[] prodNames = {"核心产品线", "互金产品线", "支付产品线"};
 
-    String[] products = {"中关村售前项目","百信售前项目","上海售前项目"};
+    String[] products = {"中关村售前项目", "百信售前项目", "上海售前项目"};
 
 
     @Autowired
     public InitDatabase(UserService userService, DepartmentService departmentService,
-                        ProductService productService,BaseProjectService baseProjectService) {
+                        ProductService productService, BaseProjectService baseProjectService) {
 
-        departmentService.deleteAll();
-        userService.deleteAll();
-        productService.deleteAll();
-        baseProjectService.deleteAll();
 
-        for (int i = 0; i < depNames.length; i++) {
-            departmentService.create(depNames[i]);
+        List users = userService.findAll();
+        if (BusiUtil.isNull(users)) {
+
+            departmentService.deleteAll();
+            userService.deleteAll();
+            productService.deleteAll();
+            baseProjectService.deleteAll();
+
+            for (int i = 0; i < depNames.length; i++) {
+                departmentService.create(depNames[i]);
+            }
+
+            for (int i = 0; i < prodNames.length; i++) {
+                productService.create(prodNames[i]);
+            }
+
+            for (int i = 0; i < products.length; i++) {
+                baseProjectService.create(products[i]);
+            }
+
+            userService.create("admin", "888888", "USER", departmentService.findAll().get(0), "技术总监", "张三");
         }
-
-        for (int i = 0; i < prodNames.length; i++) {
-            productService.create(prodNames[i]);
-        }
-
-        for (int i = 0; i < products.length; i++) {
-            baseProjectService.create(products[i]);
-        }
-
-        userService.create("admin", "888888", "USER", departmentService.findAll().get(0), "技术总监", "张三");
-
     }
 }
